web/requirements.txt
flask
python-dotenv
itsdangerous
pika